refactor: moved component mocking preview to own lib
