+++ bunit.sln
@@ -59,6 +59,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".workflows", ".workflows", 
 		.github\workflows\verification.yml = .github\workflows\verification.yml
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "bunit.web.mock", "src\bunit.web.mock\bunit.web.mock.csproj", "{3B51665F-E515-498A-8A82-BDB623089F9D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "bunit.web.mock.tests", "tests\bunit.web.mock.tests\bunit.web.mock.tests.csproj", "{050E57C2-C28F-4A9A-B35F-5725068AEF80}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -97,6 +101,14 @@ Global
 		{4CAB561E-0AFF-4516-AB9B-981F94EF1E86}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4CAB561E-0AFF-4516-AB9B-981F94EF1E86}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4CAB561E-0AFF-4516-AB9B-981F94EF1E86}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B51665F-E515-498A-8A82-BDB623089F9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B51665F-E515-498A-8A82-BDB623089F9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B51665F-E515-498A-8A82-BDB623089F9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B51665F-E515-498A-8A82-BDB623089F9D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{050E57C2-C28F-4A9A-B35F-5725068AEF80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{050E57C2-C28F-4A9A-B35F-5725068AEF80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{050E57C2-C28F-4A9A-B35F-5725068AEF80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{050E57C2-C28F-4A9A-B35F-5725068AEF80}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -111,6 +123,8 @@ Global
 		{7972A80F-30DC-4EF4-9294-7D4DD2965882} = {6EA09ED4-B714-4E6F-B0E1-4D987F8AE520}
 		{F61B8A55-3B5B-45FB-8E38-DBF9296B35C7} = {9A2B3B34-D41C-43E8-BC7D-246BEBE48D59}
 		{4CAB561E-0AFF-4516-AB9B-981F94EF1E86} = {6EA09ED4-B714-4E6F-B0E1-4D987F8AE520}
+		{3B51665F-E515-498A-8A82-BDB623089F9D} = {9A2B3B34-D41C-43E8-BC7D-246BEBE48D59}
+		{050E57C2-C28F-4A9A-B35F-5725068AEF80} = {6EA09ED4-B714-4E6F-B0E1-4D987F8AE520}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {24106918-1C86-4769-BDA6-9C80E64CD260}
