+++ src/bunit.web.mock/TestDoubles/Components/Stub{TComponent}.cs
@@ -1,4 +1,3 @@
-#if NET5_0_OR_GREATER
 using System.Collections.Generic;
 using System.Diagnostics;
 using Microsoft.AspNetCore.Components;
@@ -45,4 +44,3 @@ namespace Bunit.TestDoubles
 		}
 	}
 }
-#endif
