+++ src/bunit.web.mock/bunit.web.mock.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<RootNamespace>Bunit</RootNamespace>
+		<AssemblyName>Bunit.Web.Mock</AssemblyName>
+		<PackageVersion>0.1.1-preview</PackageVersion>
+	</PropertyGroup>
+
+	<PropertyGroup>
+		<PackageId>bunit.web.mock</PackageId>
+		<Title>bUnit.web.mock</Title>
+		<Description>
+			bUnit.web.mock contains an experimental component stubbing mocking library. This library is in preview, and the logic will likely be added back into bUnit when out of preview.
+		</Description>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\bunit.core\bunit.core.csproj" />
+		<ProjectReference Include="..\bunit.web\bunit.web.csproj" />
+	</ItemGroup>
+
+</Project>
