+++ src/bunit.web.testcomponents/bunit.web.testcomponents.csproj
@@ -24,6 +24,7 @@ NOTE: This package represents experimental features of bUnit that has been super
 		<PackageReference Include="xunit.assert" Version="2.4.1" />
 		<PackageReference Include="xunit.extensibility.core" Version="2.4.1" />
 		<PackageReference Include="xunit.extensibility.execution" Version="2.4.1" />
+		<PackageReference Include="Nerdbank.GitVersioning" Version="3.4.244" PrivateAssets="All" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
