+++ src/bunit.web/bunit.web.csproj
@@ -19,6 +19,7 @@
 		<PackageReference Include="AngleSharp.Css" Version="0.16.2" />
 		<PackageReference Include="AngleSharp.Diffing" Version="0.17.0" />
 		<PackageReference Include="AngleSharp.Wrappers" Version="2.0.0" />
+		<PackageReference Include="Nerdbank.GitVersioning" Version="3.4.244" PrivateAssets="All" />
 	</ItemGroup>
 	
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
