+++ src/bunit/bunit.csproj
@@ -25,4 +25,8 @@
 	  <ProjectReference Include="..\bunit.web\bunit.web.csproj" />
 	</ItemGroup>
 
+	<ItemGroup>
+		<PackageReference Include="Nerdbank.GitVersioning" Version="3.4.244" PrivateAssets="All" />
+	</ItemGroup>
+
 </Project>
